+++ src/reducers/index.js
@@ -12,7 +12,7 @@ const initialState = {
   placeholderOrigin: 'Choose a starting place',
   placeholderDestination: 'Choose destination',
   zoom: 16,
-  language: 'en',
+  language: null,
   compile: null,
   proximity: false,
   styles: [],